+++ core/application.go
@@ -4,7 +4,6 @@ import (
 	"errors"
 	"groups/core/model"
 	"groups/driven/corebb"
-	"groups/driven/polls"
 	"groups/driven/rewards"
 )
 
@@ -23,7 +22,6 @@ type Application struct {
 	authman       Authman
 	corebb        Core
 	rewards       Rewards
-	polls         Polls
 }
 
 // Start starts the corebb part of the application
@@ -81,10 +79,10 @@ func (app *Application) CreateUser(clientID string, id string, externalID *strin
 
 // NewApplication creates new Application
 func NewApplication(version string, build string, storage Storage, notifications Notifications, authman Authman, core *corebb.Adapter,
-	rewards *rewards.Adapter, polls *polls.Adapter, config *model.Config) *Application {
+	rewards *rewards.Adapter, config *model.Config) *Application {
 
 	application := Application{version: version, build: build, storage: storage, notifications: notifications,
-		authman: authman, corebb: core, rewards: rewards, polls: polls, config: config}
+		authman: authman, corebb: core, rewards: rewards, config: config}
 
 	//add the drivers ports/interfaces
 	application.Services = &servicesImpl{app: &application}