+++ main.go
@@ -6,7 +6,6 @@ import (
 	"groups/driven/authman"
 	"groups/driven/corebb"
 	"groups/driven/notifications"
-	"groups/driven/polls"
 	"groups/driven/rewards"
 	storage "groups/driven/storage"
 	web "groups/driver/web"
@@ -68,7 +67,7 @@ func main() {
 	}
 
 	// Instantiate a remote ServiceRegLoader to load auth service registration record from auth service
-	serviceLoader, err := authservice.NewRemoteAuthDataLoader(remoteConfig, []string{"rewards", "polls-v2"}, logs.NewLogger("groupsbb", &logs.LoggerOpts{}))
+	serviceLoader, err := authservice.NewRemoteAuthDataLoader(remoteConfig, []string{"rewards"}, logs.NewLogger("groupsbb", &logs.LoggerOpts{}))
 	if err != nil {
 		log.Fatalf("error instancing auth data loader: %s", err)
 	}
@@ -85,13 +84,6 @@ func main() {
 	}
 	rewardsAdapter := rewards.NewRewardsAdapter(rewardsServiceReg.Host, intrernalAPIKey)
 
-	// Rewards adapter
-	pollsServiceReg, err := authService.GetServiceReg("polls-v2")
-	if err != nil {
-		log.Fatalf("error finding polls service reg: %s", err)
-	}
-	pollsAdapter := polls.NewPollsAdapter(pollsServiceReg.Host, intrernalAPIKey)
-
 	config := &model.Config{
 		AuthmanAdminUINList:       authmanAdminUINList,
 		ReportAbuseRecipientEmail: notificationsReportAbuseEmail,
@@ -99,7 +91,7 @@ func main() {
 
 	//application
 	application := core.NewApplication(Version, Build, storageAdapter, notificationsAdapter, authmanAdapter,
-		coreAdapter, rewardsAdapter, pollsAdapter, config)
+		coreAdapter, rewardsAdapter, config)
 	application.Start()
 
 	//web adapter